chore: Rename the user model module name
