+++ app/modules/users/Gauth.py
@@ -10,7 +10,7 @@ from flask_jwt_extended import (
 
 # Import database models
 from models.db import db
-from models.user import Users
+from models.users import Users
 
 # Import helper function
 from app.utils.create_user_hash import create_user_hash
